+++ lib/cerbos/protobuf.rb
@@ -7,4 +7,6 @@ module Cerbos
 end
 
 require_relative "protobuf/cerbos/svc/v1/svc_services_pb"
+require_relative "protobuf/cerbos/cloud/apikey/v1/apikey_services_pb"
+require_relative "protobuf/cerbos/cloud/store/v1/store_services_pb"
 require_relative "protobuf/grpc/health/v1/health_services_pb"